Navbar improvement
navbar looks much closer to how it should, but there are still some problems. I want it to be more animated, to slowly fade in. I also want it to be responsive, as it is currently not responsive at all.

I also removed some .vue files to make it more clean.